--- web/src/components/Scrollhash/Scrollhash.js
@@ -1,36 +1,0 @@
-import { useMemo, useEffect } from 'react'
-
-import { useLocation } from '@redwoodjs/router'
-
-const Scrollhash = () => {
-  let location = useLocation()
-
-  let hashElement = useMemo(() => {
-    let hash = location.hash
-    const removeHashCharacter = (str) => {
-      const result = str.slice(1)
-      return result
-    }
-
-    if (hash) {
-      let element = document.getElementById(removeHashCharacter(hash))
-      return element
-    } else {
-      return null
-    }
-  }, [location])
-
-  useEffect(() => {
-    if (hashElement) {
-      hashElement.scrollIntoView({
-        behavior: 'smooth',
-        // block: "end",
-        inline: 'nearest',
-      })
-    }
-  }, [hashElement])
-
-  return null
-}
-
-export default Scrollhash
